+++ CHANGELOG.md
@@ -11,6 +11,9 @@ Earthquake source parameters from S-wave displacement spectra
 
 
 ## unreleased
+  - New config option `rps_from_focal_mechanism` to compute station-specific
+    S-wave radiation pattern from focal mechanism, if a focal mechanism is
+    available in the QuakeML file
   - Change default inversion algorithm to TNC (truncated
     Newton algorithm)
   - Config option `dataless` renamed to `station_metadata`